requirements.txt
requests==2.23.0
pytube3==9.6.4
opencv-python==4.2.0.32
tensorflow==2.1.0
Pillow==7.0.0
numpy==1.18.1
